tests/example-async-service/requirements.txt
grpcio>=1.14.2
grpcio-tools>=1.14.1
requests